app/requirements.txt
absl-py==0.10.0
aiofiles==0.7.0
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==0.17.0
asgiref==3.3.4
astroid==2.4.2
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.1.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.1
binaryornot==0.4.4
bleach==3.1.5
blis==0.7.4
bqplot==0.11.0
cachetools==4.1.1
catalogue==2.0.1
category-encoders==2.2.2
certifi==2020.6.20
cffi==1.14.2
cfgv==3.3.0
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.3
confuse==1.4.0
cookiecutter==1.7.2
coverage==5.4
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.2.3
dnspython==2.1.0
docopt==0.6.2
eli5==0.10.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0/en_core_web_lg-3.0.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0/en_core_web_md-3.0.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
fastapi==0.65.2
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.4.4
flatbuffers==1.12
funcy==1.16
future==0.18.2
gast==0.4.0
gensim==4.0.1
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.15
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
htmlmin==0.1.12
identify==2.2.6
idna==2.10
ImageHash==4.2.0
imageio==2.9.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.7.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.1
jinja2-time==0.2.0
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kaggle==1.5.6
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Lambdata-Jacob-Torres==0.0.1
lazy-object-proxy==1.4.3
llvmlite==0.35.0
lxml==4.6.3
Markdown==3.2.2
MarkupSafe==2.0.1
matplotlib==3.3.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mlxtend==0.18.0
murmurhash==1.0.5
nbconvert==5.6.1
nbformat==4.4.0
nbinteract==0.2.5
networkx==2.5
nodeenv==1.6.0
notebook==6.1.3
numba==0.52.0
numexpr==2.7.3
numpy==1.21.2
oauth==1.0.1
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.3.0
pandas-profiling==2.9.0
pandocfilters==1.4.2
parso==0.7.1
pathy==0.4.0
patsy==0.5.1
PDPbox==0.2.0
pep8==1.7.1
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.11.15
plotly==4.14.1
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.13.0
preshed==3.0.5
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
psutil==5.7.3
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.4
pydub==0.24.1
Pygments==2.6.1
pyLDAvis==3.3.1
pylint==2.6.0
pymongo==3.11.4
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dotenv==0.19.0
python-slugify==4.0.1
python-twitter==3.5
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
shap==0.37.0
six==1.15.0
sklearn==0.0
slicer==0.0.3
smart-open==3.0.0
sonipy==1.0
soupsieve==2.0.1
spacy==3.0.3
spacy-legacy==3.0.1
spark==0.2.1
spotipy==2.19.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
squarify==0.4.3
-e git+https://github.com/ptpt-airbnb-2/AirBnB-Optimal-Price@20abde3e951dfcbdd59a015d78305e6f2060eb1c#egg=src
srsly==2.4.0
starlette==0.14.2
statsmodels==0.12.1
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.5.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thinc==8.0.1
threadpoolctl==2.1.0
tifffile==2021.8.8
toml==0.10.2
toolz==0.11.1
tornado==6.0.4
tox==3.23.1
tqdm==4.48.2
traitlets==4.3.3
traittypes==0.2.1
translate==3.5.0
tweepy==3.10.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
uvicorn==0.14.0
virtualenv==20.2.2
virtualenv-clone==0.5.4
visions==0.5.0
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.3.0.post0
